Make th plist backend pluggable
