+++ spec/plist_helper_spec.rb
@@ -5,6 +5,7 @@ require File.expand_path('../spec_helper', __FILE__)
 module ProjectSpecs
   describe 'Xcodeproj::PlistHelper' do
     before do
+      Plist.implementation = Plist::FFI
       @plist = temporary_directory + 'plist'
     end
 
@@ -13,9 +14,9 @@ module ProjectSpecs
 
       it 'writes an XML plist file' do
         hash = { 'archiveVersion' => '1.0' }
-        DevToolsCore.stubs(:load_xcode_frameworks).returns(nil)
-        Xcodeproj.write_plist(hash, @plist)
-        result = Xcodeproj.read_plist(@plist)
+        Plist::FFI::DevToolsCore.stubs(:load_xcode_frameworks).returns(nil)
+        Plist.write_to_path(hash, @plist)
+        result = Plist.read_from_path(@plist)
         result.should == hash
         @plist.read.should.include('?xml')
       end
@@ -23,7 +24,7 @@ module ProjectSpecs
       it 'reads an ASCII plist file' do
         dir = 'Sample Project/Cocoa Application.xcodeproj/'
         path = fixture_path(dir + 'project.pbxproj')
-        result = Xcodeproj.read_plist(path)
+        result = Plist.read_from_path(path)
         result.keys.should.include?('archiveVersion')
       end
 
@@ -42,8 +43,8 @@ module ProjectSpecs
         # rubocop:enable Style/Tab
 
         hash = { 'archiveVersion' => '1.0' }
-        DevToolsCore.stubs(:load_xcode_frameworks).returns(nil)
-        Xcodeproj.write_plist(hash, @plist)
+        Plist::FFI::DevToolsCore.stubs(:load_xcode_frameworks).returns(nil)
+        Plist.write_to_path(hash, @plist)
         @plist.read.should == output
       end
     end
@@ -55,16 +56,16 @@ module ProjectSpecs
 
       it 'coerces the given path object to a string path' do
         # @plist is a Pathname
-        Xcodeproj.write_plist({}, @plist)
-        Xcodeproj.read_plist(@plist).should == {}
+        Plist.write_to_path({}, @plist)
+        Plist.read_from_path(@plist).should == {}
       end
 
       it "raises when the given path can't be coerced into a string path" do
-        lambda { Xcodeproj.write_plist({}, Object.new) }.should.raise TypeError
+        lambda { Plist.write_to_path({}, Object.new) }.should.raise TypeError
       end
 
       it "raises if the given path doesn't exist" do
-        lambda { Xcodeproj.read_plist('doesnotexist') }.should.raise Xcodeproj::Informative
+        lambda { Plist.read_from_path('doesnotexist') }.should.raise Xcodeproj::Informative
       end
 
       it 'coerces the given hash to a Hash' do
@@ -72,17 +73,17 @@ module ProjectSpecs
         def o.to_hash
           { 'from' => 'object' }
         end
-        Xcodeproj.write_plist(o, @plist)
-        Xcodeproj.read_plist(@plist).should == { 'from' => 'object' }
+        Plist.write_to_path(o, @plist)
+        Plist.read_from_path(@plist).should == { 'from' => 'object' }
       end
 
       it "raises when given a hash that can't be coerced to a Hash" do
-        lambda { Xcodeproj.write_plist(Object.new, @plist) }.should.raise TypeError
+        lambda { Plist.write_to_path(Object.new, @plist) }.should.raise TypeError
       end
 
       it 'coerces keys to strings' do
-        Xcodeproj.write_plist({ 1 => '1', :symbol => 'symbol' }, @plist)
-        Xcodeproj.read_plist(@plist).should == { '1' => '1', 'symbol' => 'symbol' }
+        Plist.write_to_path({ 1 => '1', :symbol => 'symbol' }, @plist)
+        Plist.read_from_path(@plist).should == { '1' => '1', 'symbol' => 'symbol' }
       end
 
       it 'allows hashes, strings, booleans, numbers, and arrays of hashes and strings as values' do
@@ -95,19 +96,19 @@ module ProjectSpecs
           'float' => 0.5,
           'array' => ['string in an array', { 'a hash' => 'in an array' }],
         }
-        Xcodeproj.write_plist(hash, @plist)
-        Xcodeproj.read_plist(@plist).should == hash
+        Plist.write_to_path(hash, @plist)
+        Plist.read_from_path(@plist).should == hash
       end
 
       it 'coerces values to strings if it is a disallowed type' do
-        Xcodeproj.write_plist({ '1' => 9_999_999_999_999_999_999_999_999, 'symbol' => :symbol }, @plist)
-        Xcodeproj.read_plist(@plist).should == { '1' => '9999999999999999999999999', 'symbol' => 'symbol' }
+        Plist.write_to_path({ '1' => 9_999_999_999_999_999_999_999_999, 'symbol' => :symbol }, @plist)
+        Plist.read_from_path(@plist).should == { '1' => '9999999999999999999999999', 'symbol' => 'symbol' }
       end
 
       it 'handles unicode characters in paths and strings' do
         plist = @plist.to_s + 'øµ'
-        Xcodeproj.write_plist({ 'café' => 'før yoµ' }, plist)
-        Xcodeproj.read_plist(plist).should == { 'café' => 'før yoµ' }
+        Plist.write_to_path({ 'café' => 'før yoµ' }, plist)
+        Plist.read_from_path(plist).should == { 'café' => 'før yoµ' }
       end
 
       it 'raises if a plist contains any non-supported object type' do
@@ -123,7 +124,7 @@ module ProjectSpecs
 </plist>
 EOS
         end
-        lambda { Xcodeproj.read_plist(@plist) }.should.raise TypeError
+        lambda { Plist.read_from_path(@plist) }.should.raise TypeError
       end
 
       it 'raises if a plist array value contains any non-supported object type' do
@@ -141,18 +142,18 @@ EOS
 </plist>
 EOS
         end
-        lambda { Xcodeproj.read_plist(@plist) }.should.raise TypeError
+        lambda { Plist.read_from_path(@plist) }.should.raise TypeError
       end
 
       it 'raises if for whatever reason the value could not be converted to a CFTypeRef' do
         lambda do
-          Xcodeproj.write_plist({ 'invalid' => "\xCA" }, @plist)
+          Plist.write_to_path({ 'invalid' => "\xCA" }, @plist)
         end.should.raise TypeError
       end
 
       it 'will not crash when using an empty path' do
         lambda do
-          Xcodeproj.write_plist({}, '')
+          Plist.write_to_path({}, '')
         end.should.raise IOError
       end
     end
@@ -164,27 +165,27 @@ EOS
 
       after do
         if @original_xcode_path
-          DevToolsCore.send(:remove_const, :XCODE_PATH)
-          DevToolsCore.const_set(:XCODE_PATH, @original_xcode_path)
+          Plist::FFI::DevToolsCore.send(:remove_const, :XCODE_PATH)
+          Plist::FFI::DevToolsCore.const_set(:XCODE_PATH, @original_xcode_path)
         end
       end
 
       def read_sample
         dir = 'Sample Project/Cocoa Application.xcodeproj/'
         path = fixture_path(dir + 'project.pbxproj')
-        Xcodeproj.read_plist(path)
+        Plist.read_from_path(path)
       end
 
       def stub_xcode_path(stubbed_path)
-        @original_xcode_path = DevToolsCore::XCODE_PATH
-        DevToolsCore.send(:remove_const, :XCODE_PATH)
-        DevToolsCore.const_set(:XCODE_PATH, stubbed_path)
+        @original_xcode_path = Plist::FFI::DevToolsCore::XCODE_PATH
+        Plist::FFI::DevToolsCore.send(:remove_const, :XCODE_PATH)
+        Plist::FFI::DevToolsCore.const_set(:XCODE_PATH, stubbed_path)
       end
 
       def write_temp_file_and_compare(sample)
         temp_file = File.join(SpecHelper.temporary_directory, 'out.pbxproj')
-        Xcodeproj.write_plist(sample, temp_file)
-        result = Xcodeproj.read_plist(temp_file)
+        Plist.write_to_path(sample, temp_file)
+        result = Plist.read_from_path(temp_file)
 
         sample.should == result
         File.new(temp_file).read.start_with?('<?xml').should == true
@@ -204,19 +205,19 @@ EOS
       end
 
       it 'will fallback to XML encoding if Xcode functions cannot be found' do
-        DevToolsCore.stubs(:load_xcode_frameworks).returns(Fiddle::Handle.new)
+        Plist::FFI::DevToolsCore.stubs(:load_xcode_frameworks).returns(Fiddle::Handle.new)
 
         write_temp_file_and_compare(read_sample)
       end
 
       it 'will fallback to XML encoding if Xcode methods return errors' do
-        DevToolsCore::NSData.any_instance.stubs(:writeToFileAtomically).returns(false)
+        Plist::FFI::DevToolsCore::NSData.any_instance.stubs(:writeToFileAtomically).returns(false)
 
         write_temp_file_and_compare(read_sample)
       end
 
       it 'will fallback to XML encoding if Xcode classes cannot be found' do
-        DevToolsCore::NSObject.stubs(:objc_class).returns(nil)
+        Plist::FFI::DevToolsCore::NSObject.stubs(:objc_class).returns(nil)
 
         write_temp_file_and_compare(read_sample)
       end
